requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
requests==2.32.3
python-dotenv==1.0.1
